src/bot.py
# -*- coding: utf-8 -*-
import os
import logging
from aiohttp import web
from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import Application, CommandHandler, CallbackQueryHandler, MessageHandler, filters, ContextTypes

# توکن بات اصلی
BOT_TOKEN = os.environ.get('BOT_TOKEN', '8327912063:AAEh4Q_mrVsAl9GYiSLTnQH-Cg251RxCyCY')

# آیدی نماینده
REPRESENTATIVE_ID = 96763697

# تنظیمات لاگ
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# تمام توابع اصلی بات اینجا...
async def start(update: Update, context: ContextTypes.DEFAULT_TYPE):
    keyboard = [
        [InlineKeyboardButton("معرفی کاندید", callback_data="candidate_info")],
        [InlineKeyboardButton("عکس‌ها", callback_data="photos")],
        [InlineKeyboardButton("رزومه", callback_data="resume")],
        [InlineKeyboardButton("ایده‌ها", callback_data="ideas")],
        [InlineKeyboardButton("آدرس ستادها", callback_data="addresses")],
        [InlineKeyboardButton("ارتباط با من", callback_data="contact")]
    ]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(
        "🌟 به بات کاندید خوش آمدید!\nلطفاً یکی از گزینه‌ها را انتخاب کنید:",
        reply_markup=reply_markup
    )

# 🔹 بقیه توابع رو از کد قبلی کپی کن اینجا
# [candidate_callback, photos_callback, resume_callback, ideas_callback, 
#  addresses_callback, contact_callback, button_handler, etc.]

async def handle_webhook(request):
    """دریافت پیام‌های وب‌هوک از تلگرام"""
    try:
        data = await request.json()
        update = Update.de_json(data, request.app['bot'])
        await request.app['application'].process_update(update)
        return web.Response(text="OK")
    except Exception as e:
        logger.error(f"خطا در پردازش وب‌هوک: {e}")
        return web.Response(status=400)

async def health_check(request):
    return web.Response(text="Bot is running!")

async def init_app():
    application = Application.builder().token(BOT_TOKEN).build()
    
    # اضافه کردن هندلرها
    application.add_handler(CommandHandler("start", start))
    application.add_handler(CallbackQueryHandler(button_handler))
    application.add_handler(MessageHandler(filters.TEXT & ~filters.COMMAND, text_handler))
    application.add_handler(MessageHandler(filters.PHOTO, photo_handler))
    application.add_handler(MessageHandler(filters.VOICE, voice_handler))
    application.add_error_handler(error_handler)
    
    await application.initialize()
    await application.start()
    
    app = web.Application()
    app['bot'] = application.bot
    app['application'] = application
    
    app.router.add_post(f'/{BOT_TOKEN}', handle_webhook)
    app.router.add_get('/', health_check)
    app.router.add_get('/health', health_check)
    
    return app

if __name__ == '__main__':
    PORT = int(os.environ.get('PORT', 10000))
    logger.info(f"🚀 سرور Production اجرا شد روی پورت {PORT}")
    web.run_app(init_app(), host='0.0.0.0', port=PORT)