src/requirements.txt
python-telegram-bot==21.7
aiohttp==3.9.1
python-dotenv==1.0.0